Use float32 consistently

+++ include/treelite/base.h
@@ -15,7 +15,7 @@
 namespace treelite {
 
 /*! \brief float type to be used internally */
-typedef double tl_float;
+typedef float tl_float;
 /*! \brief feature split type */
 enum class SplitFeatureType : int8_t {
   kNone, kNumerical, kCategorical
